refactor code to integrate with order and include new methods

+++ src/main/java/com/Unir/RelatosdePapel/Catalogue/exception/BookNotFoundException.java
@@ -12,7 +12,7 @@ public class BookNotFoundException extends RuntimeException {
         super(message, cause);
     }
 
-    public BookNotFoundException(Integer id) {
+    public BookNotFoundException(Long id) {
         super("Book not found with id: " + id);
     }
